LayoutTests/imported/w3c/web-platform-tests/cors/late-upload-events-expected.txt

Blocked access to external URL http://www1.localhost:8800/cors/resources/status.py?headers=custom-header
Blocked access to external URL http://www1.localhost:8800/cors/resources/status.py?headers=custom-header
Adding upload event listeners after send()


FAIL Late listeners: No preflight assert_equals: expected 200 but got 0
FAIL Late listeners: Preflight assert_equals: expected 200 but got 0
